Added Dice Class for holding numbers and keeping track of if a Die is held or not. (And if it should be re-rolled or not.)
Added ViewModel Rolls and Rules class. (Abbreviated to VM RollsRules). Where an array of Dice[] is instantiated which will be used to display the dice and be drawn from for the combination rules.

The itemcontrol in MainWindow now binds to DiceArr and displays the numbers correctly. The roll button is functional and re-rolls the numbers which update accordingly.