- Testing out different options for displaying the scoreGrid. And which one would be the easiest to adapt to display/behave like I want.
Datagrid displays correclty. ColumnTemplates would make changing the visual appearance of columns fairly easy. Only checking checkboxes doesn't update the hold buttons for some reason.

Using an itemcontrol with a datatemplate is promising. Only the rows behave individually. Making it impossible to auto size column widths which I'm not certain if it will be an issue.

Using a listview with gridview solves the column width issue. And I could edit the gridview columns to properly display checkboxes. But I do not need or want all the extra sorting and selection options.