Made the IsValid system work correctly. If no Valid option is available. The user can select any unused row for 0 point.
Implemented end of the game as well as starting a new game. Which has been tested and works accordingly.

ScoreGrid now updates not only after diceRoll but also to reflect changes when points have been noted down.

Fixed dice being able to be held at the start of the game and betwen games.